fix: remove modal wrapper
