+++ src/index.scss
@@ -5,5 +5,5 @@
 
 @import "~@edx/frontend-component-footer/dist/footer";
 
-@import './skills-builder/skills-builder-modal/skillsBuilderModal.scss';
+@import './skills-builder/skills-builder-steps/skillsBuilderSteps.scss';
 @import './skills-builder/skills-builder-header/skillsBuilderHeader.scss';
