+++ src/skills-builder/skills-builder-header/SkillsBuilderHeader.jsx
@@ -2,7 +2,7 @@ import React, { useRef } from 'react';
 import PropTypes from 'prop-types';
 import { useIntl } from '@edx/frontend-platform/i18n';
 import edXLogo from '../images/edX-logo.svg';
-import { useVisibilityFlags } from '../skills-builder-modal/view-results/data/hooks';
+import { useVisibilityFlags } from '../skills-builder-steps/view-results/data/hooks';
 import messages from './messages';
 
 const SkillsBuilderHeader = ({ isMedium }) => {
@@ -12,7 +12,6 @@ const SkillsBuilderHeader = ({ isMedium }) => {
 
   if (showSmallHeader || isMedium) {
     return (
-
       <div className="d-flex bg-primary-500 little-hero">
         <img src={edXLogo} alt="edx-logo" className="mt-2 edx-logo" />
         <div className="ml-3 vertical-line" />
@@ -26,7 +25,7 @@ const SkillsBuilderHeader = ({ isMedium }) => {
     );
   }
   return (
-    <div className="d-flex">
+    <div className="d-flex bg-primary-500">
       <img src={edXLogo} alt="edx-logo" className="mt-2 h-50" />
       <div className="ml-5 vertical-line" />
       <div className="w-100 ml-5">
