--- src/skills-builder/skills-builder-modal/index.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export { default as SkillsBuilderModal } from './SkillsBuilderModal';
