+++ src/skills-builder/skills-builder-steps/index.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
+export { default as SkillsBuilderDefault } from './SkillsBuilderDefault';
