+++ src/skills-builder/skills-builder-steps/select-preferences/CareerInterestCard.jsx
