+++ src/skills-builder/skills-builder-steps/select-preferences/GoalSelect.jsx
