+++ src/skills-builder/skills-builder-steps/select-preferences/test/SelectPreferences.test.jsx
