+++ src/skills-builder/skills-builder-steps/skillsBuilderSteps.scss
