+++ src/skills-builder/skills-builder-steps/view-results/ProductCardGrid.jsx
