+++ src/skills-builder/skills-builder-steps/view-results/ProductTypeBanner.jsx
