+++ src/skills-builder/skills-builder-steps/view-results/RecommendationCard.jsx
