+++ src/skills-builder/skills-builder-steps/view-results/RecommendationStack.jsx
