+++ src/skills-builder/skills-builder-steps/view-results/RelatedSkillsInteractiveBoxSet.jsx
