+++ src/skills-builder/skills-builder-steps/view-results/data/constants.js
